requirements.txt
livekit-agents[sarvam,openai,google,silero,turn-detector]>=1.0.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv>=1.0.0
ultralytics 
sentence_transformers
transformers
timm
qdrant-client
openai
langchain-community
langchain-text-splitters
pypdf